Module Capnp_rpc_unix.Cap_file

Utilities for capability files.

val load : Vat.t -> string -> (_ Capnp_rpc_lwt.Sturdy_ref.t, [> `Msg of string ]) Stdlib.result

load vat path loads the contents of path as a capability URI.

val save_sturdy : Vat.t -> _ Capnp_rpc_lwt.Sturdy_ref.t -> string -> (unit, [> `Msg of string ]) Stdlib.result

save_sturdy vat sr path saves sr to path, with a mode of 0o600.

val save_service : Vat.t -> Capnp_rpc_net.Restorer.Id.t -> string -> (unit, [> `Msg of string ]) Stdlib.result

save_service vat id path saves vat/id to path, with a mode of 0o600.